requirements.txt
cookiecutter==1.7.2
neptune-client==0.4.130
numpy==1.19.4
pandas==0.20.3
python-decouple==3.3
pytorch-lightning==1.0.7
torch==1.7.0
transformers==3.5.1
